requirements.txt
flask>=3.0.0
requests>=2.31.0
beautifulsoup4>=4.12.3
google-genai>=1.2.0
duckduckgo-search>=5.3.0
rembg>=2.0.57
pillow>=10.2.0
setuptools>=69.1.1
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
